+++ src/main/java/app/domain/models/LoanDetail.java
@@ -1,16 +1,18 @@
 package app.domain.models;
 
+import java.math.BigDecimal;
+
 public class LoanDetail {
 
-    private double approvedAmount;
-    private double interestRate;
+    private BigDecimal approvedAmount;
+    private BigDecimal interestRate;
     private String previousState;
     private String newState;
     private String analystId;
 
     // ── Constructor ───────────────────────────────────────────────────
 
-    public LoanDetail(double approvedAmount, double interestRate,
+    public LoanDetail(BigDecimal approvedAmount, BigDecimal interestRate,
                        String previousState, String newState, String analystId) {
         this.approvedAmount = approvedAmount;
         this.interestRate = interestRate;
@@ -21,11 +23,11 @@ public class LoanDetail {
 
     // ── Getters & Setters ─────────────────────────────────────────────
 
-    public double getApprovedAmount() { return approvedAmount; }
-    public void setApprovedAmount(double approvedAmount) { this.approvedAmount = approvedAmount; }
+    public BigDecimal getApprovedAmount() { return approvedAmount; }
+    public void setApprovedAmount(BigDecimal approvedAmount) { this.approvedAmount = approvedAmount; }
 
-    public double getInterestRate() { return interestRate; }
-    public void setInterestRate(double interestRate) { this.interestRate = interestRate; }
+    public BigDecimal getInterestRate() { return interestRate; }
+    public void setInterestRate(BigDecimal interestRate) { this.interestRate = interestRate; }
 
     public String getPreviousState() { return previousState; }
     public void setPreviousState(String previousState) { this.previousState = previousState; }
